# internal/docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API to track performance",
        "title": "Time Tracker",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/users": {
            "get": {
                "description": "Get a list of users based on filters and pagination",
                "produces": [
                    "application/json"
                ],
                "summary": "Get users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User filters",
                        "name": "filters",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "User pagination",
                        "name": "pagination",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of users",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1_user.getUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new user with the provided passport number and series",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1_user.createUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1_user.createUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/{userID}": {
            "delete": {
                "description": "Delete a user with the provided user ID",
                "summary": "Delete a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to delete",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update a user with the provided user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID to update",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update user request",
                        "name": "updateUserRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1_user.updateUsersRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated successfully",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1_user.updateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/{userID}/task-summaries": {
            "get": {
                "description": "Get the task summaries for a specific user within a given period",
                "produces": [
                    "application/json"
                ],
                "summary": "Get task summaries by user and period",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "Start of the period",
                        "name": "periodStart",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date-time",
                        "description": "End of the period",
                        "name": "periodEnd",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response",
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1_task.getUserTaskSummariesByPeriodResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    }
                }
            }
        },
        "/users/{userID}/tasks/{taskID}/status": {
            "put": {
                "description": "Set the status of a task for a specific user",
                "consumes": [
                    "application/json"
                ],
                "summary": "Set task status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "userID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "taskID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task status request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_controllers_v1_task.changeTaskStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/time-tracker_internal_controllers_utils_http.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_controllers_v1_task.changeTaskStatusRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "$ref": "#/definitions/time-tracker_internal_domain.TaskStatus"
                }
            }
        },
        "internal_controllers_v1_task.getUserTaskSummariesByPeriodResponse": {
            "type": "object",
            "properties": {
                "taskSummaries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_controllers_v1_task.taskSummary"
                    }
                }
            }
        },
        "internal_controllers_v1_task.taskSummary": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "totalInWorkDuration": {
                    "type": "string"
                }
            }
        },
        "internal_controllers_v1_user.createUserRequest": {
            "type": "object",
            "required": [
                "passportNumber"
            ],
            "properties": {
                "passportNumber": {
                    "type": "string"
                }
            }
        },
        "internal_controllers_v1_user.createUserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/internal_controllers_v1_user.user"
                }
            }
        },
        "internal_controllers_v1_user.getUsersResponse": {
            "type": "object",
            "properties": {
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_controllers_v1_user.user"
                    }
                }
            }
        },
        "internal_controllers_v1_user.updateUserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/internal_controllers_v1_user.user"
                }
            }
        },
        "internal_controllers_v1_user.updateUsersRequest": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "properties": {
                        "address": {
                            "type": "string"
                        },
                        "name": {
                            "type": "string"
                        },
                        "passportSeriesAndNumber": {
                            "type": "string"
                        },
                        "patronymic": {
                            "type": "string"
                        },
                        "surname": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "internal_controllers_v1_user.user": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "passportNumber": {
                    "type": "string"
                },
                "patronymic": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "time-tracker_internal_controllers_utils_http.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "time-tracker_internal_domain.TaskStatus": {
            "type": "string",
            "enum": [
                "iddle",
                "in-work"
            ],
            "x-enum-varnames": [
                "TaskStatusIddle",
                "TaskStatusInWork"
            ]
        }
    }
}